Multiply fee tiers 10 times
* Add a new fee tier: 0.008%
* Multiply fee tiers 10 times
* Add FEE_UNITS = 100,000
* Update tests

+++ contracts/mock/MockEncoder.sol
@@ -6,7 +6,7 @@ import {Router} from '../periphery/Router.sol';
 contract MockEncoder {
   function encodeSwapCallbackData(
     address tokenIn,
-    uint16 fee,
+    uint24 fee,
     address tokenOut
   ) external view returns (bytes memory) {
     return
